rename plugin to shoalsoft-pants-mcp-plugin
