+++ src/python/shoalsoft/pants_mcp_plugin/pants_integration_testutil.py
