+++ src/python/shoalsoft/pants_mcp_plugin/requirements.txt
